Add files via upload
This project is called Auto-Mailer .
It's primary purpose is to facilitate seamless e-mailing via Google Mail. In this Script, the user only needs to enter the recipient of the mail, the Subject of the mail, and the Content of the Mail itself. The rest is handled by the script. It uses NodeJS with the puppeteer module to facilitate automated browser usage which first accesses www.gmail.com , then proceeds to log in to Gmail using the credentials provided to it beforehand in a separate file. Once logged in, the script tells the browser to create a new mail, and then enters the details of the mail that it took as input earlier. Once all the data has been inputted, the browser sends the mail and shows a pop up, letting us know that the mail has been sent.